requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.42.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-api-core==2.17.1
google-auth==2.28.1
gunicorn